refactor: Remove ESLint configuration and add core chess components
- Deleted the ESLint configuration file.
- Updated import paths in chessalyzer.ts to follow a consistent naming convention.
- Introduced new core components: chess-board.ts, chess-worker.ts, game-parser.ts, game-processor.ts, and worker-pool.ts for improved game processing and multithreading capabilities.
- Added base tracker classes and specific trackers for game and piece analysis.
- Implemented heatmap tracking for pieces and tiles to enhance game analysis features.
