--- eslint.config.mjs
@@ -1,28 +1,0 @@
-// @ts-check
-
-import eslint from '@eslint/js';
-import tseslint from 'typescript-eslint';
-
-export default tseslint.config(
-	eslint.configs.recommended,
-	tseslint.configs.recommendedTypeChecked,
-	{
-		languageOptions: {
-			parserOptions: {
-				projectService: true,
-				tsconfigRootDir: import.meta.dirname
-			}
-		}
-	},
-	{
-		ignores: ['*', '!src/']
-	},
-	{
-		rules: {
-			'@typescript-eslint/no-unused-vars': [
-				'error',
-				{ argsIgnorePattern: '^_', varsIgnorePattern: '^_' }
-			]
-		}
-	}
-);
