+++ src/js/chessalyzer.ts
@@ -1,7 +1,7 @@
-import Chessalyzer from './core/Chessalyzer.js';
-import PieceTracker from './tracker/PieceTrackerBase.js';
-import TileTracker from './tracker/TileTrackerBase.js';
-import GameTracker from './tracker/GameTrackerBase.js';
-import BaseTracker from './tracker/BaseTracker.js';
+import Chessalyzer from './core/chessalyzer.js';
+import PieceTracker from './tracker/piece-tracker-base.js';
+import TileTracker from './tracker/tile-tracker-base.js';
+import GameTracker from './tracker/game-tracker-base.js';
+import BaseTracker from './tracker/base-tracker.js';
 
 export { Chessalyzer, PieceTracker, GameTracker, TileTracker, BaseTracker };
