+++ src/js/core/worker-pool.ts
